requirements.txt
langchain==0.3.25
langchain-openai==0.3.16
langchain-pinecone==0.2.4
langchain-community==0.3.24
langgraph==0.4.1
tavily-python==0.5.1
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
pydantic==2.11.3
pinecone==6.0.0
ragas==0.2.14
beautifulsoup4==4.13.3
httpx==0.28.1
langsmith==0.3.45
sse-starlette==2.3.5
